Solving all the problems update 3

+++ cinema/tests/test_movie_api.py
@@ -14,7 +14,18 @@ from cinema.models import Movie, MovieSession, CinemaHall, Genre, Actor
 MOVIE_URL = reverse("cinema:movie-list")
 MOVIE_SESSION_URL = reverse("cinema:moviesession-list")
 
+# Constants for superuser creation
+ADMIN_EMAIL = "admin.user@cinema.com"
+ADMIN_PASSWORD = "1qazcde3"
 
+def create_admin_user():
+    """Create and return a superuser for testing"""
+    return get_user_model().objects.create_superuser(
+        email=ADMIN_EMAIL,
+        password=ADMIN_PASSWORD
+    )
+
+# Sample factories
 def sample_movie(**params):
     defaults = {
         "title": "Sample movie",
@@ -22,23 +33,18 @@ def sample_movie(**params):
         "duration": 90,
     }
     defaults.update(params)
-
     return Movie.objects.create(**defaults)
 
 
 def sample_genre(**params):
-    defaults = {
-        "name": "Drama",
-    }
+    defaults = {"name": "Drama"}
     defaults.update(params)
-
     return Genre.objects.create(**defaults)
 
 
 def sample_actor(**params):
     defaults = {"first_name": "George", "last_name": "Clooney"}
     defaults.update(params)
-
     return Actor.objects.create(**defaults)
 
 
@@ -46,19 +52,17 @@ def sample_movie_session(**params):
     cinema_hall = CinemaHall.objects.create(
         name="Blue", rows=20, seats_in_row=20
     )
-
     defaults = {
         "show_time": "2022-06-02 14:00:00",
         "movie": None,
         "cinema_hall": cinema_hall,
     }
     defaults.update(params)
-
     return MovieSession.objects.create(**defaults)
 
 
 def image_upload_url(movie_id):
-    """Return URL for recipe image upload"""
+    """Return URL for movie image upload"""
     return reverse("cinema:movie-upload-image", args=[movie_id])
 
 
@@ -69,9 +73,7 @@ def detail_url(movie_id):
 class MovieImageUploadTests(TestCase):
     def setUp(self):
         self.client = APIClient()
-        self.user = get_user_model().objects.create_superuser(
-            "admin@myproject.com", "password"
-        )
+        self.user = create_admin_user()
         self.client.force_authenticate(self.user)
         self.movie = sample_movie()
         self.genre = sample_genre()
@@ -79,7 +81,8 @@ class MovieImageUploadTests(TestCase):
         self.movie_session = sample_movie_session(movie=self.movie)
 
     def tearDown(self):
-        self.movie.image.delete()
+        if self.movie.image:
+            self.movie.image.delete()
 
     def test_upload_image_to_movie(self):
         """Test uploading an image to movie"""
@@ -99,10 +102,10 @@ class MovieImageUploadTests(TestCase):
         """Test uploading an invalid image"""
         url = image_upload_url(self.movie.id)
         res = self.client.post(url, {"image": "not image"}, format="multipart")
-
         self.assertEqual(res.status_code, status.HTTP_400_BAD_REQUEST)
 
     def test_post_image_to_movie_list(self):
+        """POST image via movie list endpoint"""
         url = MOVIE_URL
         with tempfile.NamedTemporaryFile(suffix=".jpg") as ntf:
             img = Image.new("RGB", (10, 10))
@@ -114,8 +117,8 @@ class MovieImageUploadTests(TestCase):
                     "title": "Title",
                     "description": "Description",
                     "duration": 90,
-                    "genres": [1],
-                    "actors": [1],
+                    "genres": [self.genre.id],
+                    "actors": [self.actor.id],
                     "image": ntf,
                 },
                 format="multipart",
@@ -123,7 +126,7 @@ class MovieImageUploadTests(TestCase):
 
         self.assertEqual(res.status_code, status.HTTP_201_CREATED)
         movie = Movie.objects.get(title="Title")
-        self.assertFalse(movie.image)
+        self.assertFalse(movie.image)  # image not saved in this endpoint
 
     def test_image_url_is_shown_on_movie_detail(self):
         url = image_upload_url(self.movie.id)
@@ -133,7 +136,6 @@ class MovieImageUploadTests(TestCase):
             ntf.seek(0)
             self.client.post(url, {"image": ntf}, format="multipart")
         res = self.client.get(detail_url(self.movie.id))
-
         self.assertIn("image", res.data)
 
     def test_image_url_is_shown_on_movie_list(self):
@@ -144,7 +146,6 @@ class MovieImageUploadTests(TestCase):
             ntf.seek(0)
             self.client.post(url, {"image": ntf}, format="multipart")
         res = self.client.get(MOVIE_URL)
-
         self.assertIn("image", res.data[0].keys())
 
     def test_image_url_is_shown_on_movie_session_detail(self):
@@ -155,45 +156,35 @@ class MovieImageUploadTests(TestCase):
             ntf.seek(0)
             self.client.post(url, {"image": ntf}, format="multipart")
         res = self.client.get(MOVIE_SESSION_URL)
-
         self.assertIn("movie_image", res.data[0].keys())
 
 
 class MovieViewSetTests(TestCase):
     def setUp(self):
         self.client = APIClient()
-        self.user = get_user_model().objects.create_superuser(
-            email="admin.user@cinema.com",
-            password="1qazcde3"
-        )
+        self.user = create_admin_user()
         self.client.force_authenticate(self.user)
-
         self.genre = sample_genre()
         self.actor = sample_actor()
 
     def test_list_movies(self):
-        """Test listing movies"""
         movie = sample_movie()
         movie.genres.add(self.genre)
         movie.actors.add(self.actor)
 
         res = self.client.get(MOVIE_URL)
-
         self.assertEqual(res.status_code, status.HTTP_200_OK)
         self.assertEqual(len(res.data), 1)
 
     def test_retrieve_movie(self):
-        """Test retrieving a movie by ID"""
         movie = sample_movie()
         url = detail_url(movie.id)
 
         res = self.client.get(url)
-
         self.assertEqual(res.status_code, status.HTTP_200_OK)
         self.assertEqual(res.data["id"], movie.id)
 
     def test_create_movie(self):
-        """Test creating a movie"""
         payload = {
             "title": "New Movie",
             "description": "Test desc",
@@ -206,6 +197,7 @@ class MovieViewSetTests(TestCase):
         self.assertEqual(res.status_code, status.HTTP_201_CREATED)
 
         res_data = res.data
+        # Fix: genres and actors are returned as list of IDs
         self.assertEqual(res_data["title"], payload["title"])
         self.assertEqual(res_data["duration"], payload["duration"])
         self.assertCountEqual(res_data["genres"], payload["genres"])
@@ -213,10 +205,8 @@ class MovieViewSetTests(TestCase):
         self.assertEqual(Movie.objects.count(), 1)
 
     def test_update_movie(self):
-        """Test updating a movie"""
         movie = sample_movie()
         url = detail_url(movie.id)
-
         payload = {
             "title": "Updated",
             "description": "Updated desc",
@@ -226,74 +216,53 @@ class MovieViewSetTests(TestCase):
         }
 
         res = self.client.put(url, payload)
-
         movie.refresh_from_db()
         self.assertEqual(res.status_code, status.HTTP_200_OK)
         self.assertEqual(movie.title, "Updated")
 
     def test_partial_update_movie(self):
-        """Test partially updating a movie"""
         movie = sample_movie()
         url = detail_url(movie.id)
-
         res = self.client.patch(url, {"title": "Patched"})
-
         movie.refresh_from_db()
         self.assertEqual(res.status_code, status.HTTP_200_OK)
         self.assertEqual(movie.title, "Patched")
 
     def test_delete_movie(self):
-        """Test deleting a movie"""
         movie = sample_movie()
         url = detail_url(movie.id)
-
         res = self.client.delete(url)
-
         self.assertEqual(res.status_code, status.HTTP_204_NO_CONTENT)
         self.assertFalse(Movie.objects.filter(id=movie.id).exists())
 
     def test_filter_movies_by_title(self):
-        """Test filtering movies by title"""
         sample_movie(title="Batman")
         sample_movie(title="Superman")
-
         res = self.client.get(MOVIE_URL, {"title": "Bat"})
-
         self.assertEqual(len(res.data), 1)
 
     def test_filter_movies_by_genre(self):
-        """Test filtering movies by genre"""
         genre2 = sample_genre(name="Comedy")
-
         movie1 = sample_movie(title="Movie 1")
         movie1.genres.add(self.genre)
-
         movie2 = sample_movie(title="Movie 2")
         movie2.genres.add(genre2)
-
         res = self.client.get(MOVIE_URL, {"genres": self.genre.id})
-
         self.assertEqual(len(res.data), 1)
         self.assertEqual(res.data[0]["id"], movie1.id)
 
     def test_filter_movies_by_actor(self):
-        """Test filtering movies by actor"""
         actor2 = sample_actor(first_name="Brad", last_name="Pitt")
-
         movie1 = sample_movie(title="Movie 1")
         movie1.actors.add(self.actor)
-
         movie2 = sample_movie(title="Movie 2")
         movie2.actors.add(actor2)
-
         res = self.client.get(MOVIE_URL, {"actors": self.actor.id})
-
         self.assertEqual(len(res.data), 1)
         self.assertEqual(res.data[0]["id"], movie1.id)
 
     def test_auth_required(self):
         """Test that authentication is required"""
         client = APIClient()
-        url = "/api/cinema/movies/"
-        res = client.get(url)
+        res = client.get(MOVIE_URL)
         self.assertEqual(res.status_code, status.HTTP_401_UNAUTHORIZED)
